fix: 투표 목록 조회 응답에 댓글 수 추가 (#241)
