+++ src/main/java/com/moa/moa_server/domain/comment/repository/CommentRepository.java
@@ -1,6 +1,7 @@
 package com.moa.moa_server.domain.comment.repository;
 
 import com.moa.moa_server.domain.comment.entity.Comment;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -15,4 +16,7 @@ public interface CommentRepository extends JpaRepository<Comment, Long>, Comment
   @Modifying
   @Query("UPDATE Comment c SET c.deletedAt = CURRENT_TIMESTAMP WHERE c.vote.id = :voteId")
   void softDeleteByVoteId(@Param("voteId") Long voteId);
+
+  @Query("SELECT c.vote.id, COUNT(c) FROM Comment c WHERE c.vote.id IN :voteIds GROUP BY c.vote.id")
+  List<Object[]> countCommentsByVoteIds(@Param("voteIds") List<Long> voteIds);
 }
